% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delong.alt.R
\name{auc.index.linearize}
\alias{auc.index.linearize}
\title{An asymptotically equivalent linear representation of the index AUC}
\usage{
auc.index.linearize(w, d, beta, infl.fn, deriv.fn = NULL)
}
\arguments{
\item{w}{matrix of observed covariates, one vector of covariate per
row}

\item{d}{vector of 0s and 1s, corresponding to units' control/case
status}

\item{beta}{the estimated coefficient vector for the index AUC}

\item{infl.fn}{the influence function for beta; see details for
specification}

\item{deriv.fn}{NULL or the derivative of the AUC at the true beta; see
details for specification}
}
\value{
a vector consisting of the terms of the linearization, as
many terms as there are units (==length(w))
}
\description{
This routine computes the linearization of the index AUC
described in the method section of the referenced
manuscript.
}
\details{
Besides the data (W,D) and estimated coefficient vector, the
parameters include two functions. The parameter infl.fn computes an
influence function for the estimate of the coefficient vector. The
only obligatory parameters are the data, W and D. The parameter
deriv.fn returns the derivative of the AUC at a point beta. Its
obligatory parameters are the data, W and D, and the point beta. If
deriv.fn==NULL the derivative is numerically estimated using
numDeriv::grad.

The simulation and data analysis routines supplied with the
referenced manuscript give examples of the use of this routine for
nonparametric inference on the difference of AUCs. Also supplied
there are examples of infl.fn and deriv.fn for various models such
as LDA and probit regression.
}
\seealso{
Alexis Doyle-Connolly and Haben Michael, Nonparametric Estimation of the  AUC of an Index with Estimated Parameters (2023)
}
